Linux cleanup.

+++ reaper_drivenbymoss/MarkerProcessor.cpp
@@ -44,7 +44,7 @@ void MarkerProcessor::Process(std::string command, std::deque<std::string> &path
 	const char *cmd = path.at(1).c_str();
 
 	const std::vector<int> markers = Marker::GetMarkers(project);
-	if (index < 0 || index >= markers.size())
+	if (index < 0 || index >= (int)markers.size())
 		return;
 	const int markerID = markers.at(index);
